requirements.txt
dj-database-url==0.5.0
django-ace==1.0.11
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.18.4
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django==4.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2021.3
redis==4.1.0
requests==2.27.1
whitenoise==5.3.0
